address review comment to move the code to seperate file make the background loop to check and clear function tables
