+++ lib/Common/Memory/CMakeLists.txt
@@ -3,6 +3,7 @@ set (CCM_SOURCE_FILES ${CCM_SOURCE_FILES}
     ArenaAllocator.cpp
     CustomHeap.cpp
     CommonMemoryPch.cpp
+    DelayDeletingFunctionTable.cpp
     EtwMemoryTracking.cpp
     ForcedMemoryConstraints.cpp
     HeapAllocator.cpp
